commands: move untrack code to untrack.rs
